# ENH: stats.invgauss: use Boost implementation of ppf/isf