#### Reference issue
Supersedes gh-13665

#### What does this implement/fix?
gh-13665 implemented `ppf` and `isf` for `invgauss`, but we decided that using Boost was a better option.

This replaces our `invgauss` `ppf` and `isf` methods with those from Boost. It's faster and more accurate than main and gh-13665.

#### Additional information
I tested a62f16c2f98e5c9024d2ff7c38e1f11ccda7ee09 against main with an experiment like this.

<details>

```
import numpy as np
from scipy import stats

rng = np.random.default_rng(6424058033727397440)
qrng = stats.qmc.Halton(3, seed=rng)
sample = qrng.random(1000)
scaled = stats.qmc.scale(sample, [-8, -1, -3], [np.log10(0.5), 1, 3]) 
pz, spz, az = scaled.T
pz = 10**pz
pz[spz < 0] = 1 - pz[spz < 0]
az = 10**az
# az = np.logspace(-3, 3, 11)
# pz = np.logspace(-8, -1, 8)
# pz = np.concatenate([pz, [0.5], 1-pz[::-1]])[:, np.newaxis]

def relative_error(res, ref):
    return np.float64(abs((res-ref)/ref))

xz = stats.invgauss.ppf(pz, az)
res = stats.invgauss.cdf(xz, az)
err = relative_error(res, pz)
```

</details>

Time to evaluate `ppf` of those 1000 points:
Main: 2.08 s ± 10.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
PR: 2.18 ms ± 11.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
(similarly with isf)

Here were the error histograms:
![image](https://user-images.githubusercontent.com/6570539/183221950-9e019af9-771d-4d77-9910-c296e90619c7.png)

That little orange bar at the right concerned me, and I found that `ppf` was very inaccurate when `x` was above `0.5` and `mu` was high (>35); the same was true of `sf`. But both were accurate when `x` was below `0.5`, though, so I added the condition that performs the evaluation in the correct tail.

Also, sometimes Boost's `ppf` and `isf` produced NaNs. In these cases, we resort to the default implementation.

After the modifications in 9bdf2513e22956ee054f1d599b9971344d4ecac7, there are no NaNs, and the maximum error is `3.200609142735203e-10`. The geometric mean of the nonzero errors is `2.4529612751351497e-15`. Ignoring the warnings produced by Boost seems safe, given the measures to replace inaccurate entries. (And yes, I think replacing is preferable to `_lazywhere` because it's so fast anyway...)