@ursk 2024-03-26
This change produces wildly incorrect results under specific circumstances:

```
from scipy import stats
stats.invgauss(mu=2./2.001, scale=2.001).ppf(0.9643504) 
```

Expected output (what Scipy 1.9.3 produces): `6.656304078844659`
Observed output: `8.954132303260685e+248`

This broke a test in Tensorflow Probability [here](https://github.com/tensorflow/probability/blob/main/tensorflow_probability/python/distributions/inverse_gaussian_test.py).


I'm not sure whether this is related to "That little orange bar at the right", or if it exposed a different edge case.